test(memory): Move LongResend test into test/memory. Add additional tests. Assert memory usage. Run in CI.
