+++ package.json
@@ -48,6 +48,7 @@
     "test-integration-no-resend": "jest --forceExit --testTimeout=10000 --testPathIgnorePatterns='resend|Resend' --testNamePattern='^((?!(resend|Resend|resent|Resent|gap|Gap)).)*$' test/integration/*.test.*",
     "test-integration-resend": "jest --forceExit --testTimeout=15000 --testNamePattern='(resend|Resend|resent|Resent)' test/integration/*.test.*",
     "test-integration-dataunions": "jest --forceExit --testTimeout=15000 --runInBand test/integration/dataunion",
+    "test-memory": "node --gc-global --predictable-gc-schedule node_modules/.bin/jest --forceExit --testTimeout=15000 --runInBand test/integration/dataunion",
     "test-flakey": "jest --forceExit test/flakey/*",
     "test-browser": "node ./test/browser/server.js & node node_modules/nightwatch/bin/nightwatch ./test/browser/browser.js && pkill -f server.js",
     "install-example": "cd examples/webpack && npm ci",
